+++ app/src/main/java/com/altruist/ui/screens/register/RegisterScreen2.kt
@@ -1,47 +1,36 @@
-package com.altruist.ui.screens
+package com.altruist.ui.screens.register
 
-import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
-import androidx.compose.foundation.border
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.foundation.text.KeyboardOptions
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.ArrowDropDown
 import androidx.compose.material3.*
 import androidx.compose.runtime.*
 import androidx.compose.ui.*
 import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.text.input.PasswordVisualTransformation
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.altruist.R
-import com.altruist.ui.components.AltruistTextField
+import com.altruist.ui.components.AltruistLabeledTextField
 import com.altruist.ui.components.AltruistSnackbarHost
 import com.altruist.ui.theme.BackgroundTop
 import com.altruist.ui.theme.BackgroundBottom
-import com.altruist.ui.components.PrimaryButton
 import com.altruist.ui.components.SecondaryButton
-import com.altruist.ui.theme.TitleMediumTextStyle
 import com.altruist.viewmodel.RegisterViewModel
 
 @Composable
-fun RegisterScreen3(
+fun RegisterScreen2(
     viewModel: RegisterViewModel = hiltViewModel(),
-    onRegister3Success: () -> Unit
+    onRegister2Success: () -> Unit
 ) {
-    val verificationCode by viewModel.verificationCode.collectAsState()
-    val inputCode by viewModel.inputCode.collectAsState()
+    val email by viewModel.email.collectAsState()
+    val password by viewModel.password.collectAsState()
+    val repeatPassword by viewModel.repeatPassword.collectAsState()
     val errorMessage by viewModel.errorMessage.collectAsState()
-
-    val register3Success by viewModel.register3Success.collectAsState()
+    val register2Success by viewModel.register2Success.collectAsState()
+    val isLoading by viewModel.isLoading.collectAsState()
 
     val snackbarHostState = remember { SnackbarHostState() }
 
@@ -56,9 +45,10 @@ fun RegisterScreen3(
         }
     }
 
-    LaunchedEffect(register3Success) {
-        if (register3Success){
-            onRegister3Success()
+    LaunchedEffect(register2Success) {
+        if (register2Success) {
+            onRegister2Success()
+            viewModel.resetCodigoEnviado() // para evitar navegación doble
         }
     }
 
@@ -103,33 +93,40 @@ fun RegisterScreen3(
                 Spacer(modifier = Modifier.weight(0.05f))
 
                 Column(
-                    modifier = Modifier
-                        .padding(horizontal = 40.dp),
-                    verticalArrangement = Arrangement.spacedBy(60.dp),
-                    horizontalAlignment = Alignment.CenterHorizontally
+                    verticalArrangement = Arrangement.spacedBy(24.dp),
+                    modifier = Modifier.fillMaxWidth()
                 ) {
-                    Text(
-                        text = "Por favor, introduce el código que hemos enviado a tu correo",
-                        style = TitleMediumTextStyle,
-                        textAlign = TextAlign.Center,
-                        modifier = Modifier.padding(horizontal = 16.dp)
+                    AltruistLabeledTextField(
+                        label = "E-mail",
+                        value = email,
+                        onValueChange = viewModel::onEmailChange,
+                        placeholder = "E-mail"
                     )
 
-                    AltruistTextField(
-                        value = inputCode,
-                        onValueChange = viewModel::onInputCodeChange,
-                        placeholder = "Código de verificación",
-                        textAlign = TextAlign.Center
+                    AltruistLabeledTextField(
+                        label = "Contraseña",
+                        value = password,
+                        onValueChange = viewModel::onPasswordChange,
+                        placeholder = "Contraseña",
+                        visualTransformation = PasswordVisualTransformation()
+                    )
+
+                    AltruistLabeledTextField(
+                        label = "Repita su contraseña",
+                        value = repeatPassword,
+                        onValueChange = viewModel::onRepeatPasswordChange,
+                        placeholder = "Repita su contraseña",
+                        visualTransformation = PasswordVisualTransformation()
                     )
                 }
 
                 Spacer(modifier = Modifier.weight(0.15f))
 
                 SecondaryButton(
-                    text = "Continuar",
-                    onClick = {viewModel.onContinueFromRegister3Click()},
+                    text = if (isLoading) "Cargando..." else "Continuar",
+                    onClick = {viewModel.onContinueFromRegister2Click()},
                     modifier = Modifier.fillMaxWidth(),
-                    enabled = inputCode.length == 6
+                    enabled = !isLoading
                 )
 
                 Spacer(modifier = Modifier.weight(0.15f))
